refactor testpage to login
